Test : HP 시스템 코드 리팩토링중. 누가 죽였는지 알기 위해.

+++ Assets/Scripts/Data/PlayerInGameData.cs
@@ -10,7 +10,7 @@ public struct PlayerInGameData : IEquatable<PlayerInGameData>, INetworkSerializa
     //public PlayerAttackAnimState playerAttackAnimState;
     public PlayerGameState playerGameState;
     public FixedString64Bytes playerName;
-    public sbyte score;
+    public int score;
     public Character characterClass;
     public sbyte hp;
     public sbyte maxHp;
